chore: moved the upload button to the dashboard, instead of the register screen.
